+++ test/context/services/context-get-default.service.test.ts
@@ -14,7 +14,6 @@ describe('ContextGetDefaultService', () => {
         name: 'castle-node',
         version,
       },
-      client_id: 'client_id',
       ip: '1.2.3.4',
     };
 
@@ -29,17 +28,12 @@ describe('ContextGetDefaultService', () => {
       headers: {
         'x-forwarded-for': '1.2.3.4',
         'x-castle-client-id': 'client_id',
-        cookies: 'client_id',
       },
       body: {},
     } as ExpressRequest;
 
     it('generates default context', () => {
-      const received = ContextGetDefaultService.call(
-        mockRequest,
-        undefined,
-        config
-      );
+      const received = ContextGetDefaultService.call(mockRequest, config);
       expect(received).toMatchObject(expected);
     });
   });