+++ test/context/services/context-prepare.service.test.ts
@@ -22,7 +22,6 @@ describe('ContextPrepareService', () => {
     });
 
     const options = {
-      cookies: '__cid=abcd;',
       context: {
         client_id: 'client_id',
         active: true,